reorganized files, deleted unused files
