+++ frontend/components/learning_circle_form/HelpContainer.jsx
@@ -1,6 +1,7 @@
 import React from 'react'
 import { Step1, Step2, Step3, Step4, Step5 } from './helpText'
 
+
 export default class CreateLearningCircleForm extends React.Component{
   constructor(props){
     super(props);
