+++ frontend/components/learning_circle_form/course_search/Filter.jsx
@@ -1,5 +1,6 @@
 import React from 'react'
 
+
 const Filter = ({filter, active, updateActiveFilter}) => {
   const filterNames = {
     'location': 'Location',
