+++ frontend/components/learning_circle_form/course_search/FilterForm.jsx
@@ -1,8 +1,7 @@
 import React, { Component } from 'react'
-import LocationFilterForm from './LocationFilterForm'
 import TopicsFilterForm from './TopicsFilterForm'
-import MeetingDaysFilterForm from './MeetingDaysFilterForm'
 import OrderCoursesForm from './OrderCoursesForm'
+
 
 const FilterForm = (props) => {
   const closeFilter = () => { props.updateActiveFilter(null) };
@@ -10,12 +9,8 @@ const FilterForm = (props) => {
 
   const internalForm = () => {
     switch (props.activeFilter) {
-      case 'location':
-      return <LocationFilterForm {...props} closeFilter={closeFilter} />;
       case 'topics':
       return <TopicsFilterForm {...props} />;
-      case 'meetingDays':
-      return <MeetingDaysFilterForm { ...props} />;
       case 'orderCourses':
       return <OrderCoursesForm { ...props} />;
     }
