+++ frontend/components/learning_circle_form/course_search/Search.jsx
@@ -6,6 +6,7 @@ import { SEARCH_PROPS } from '../../../constants'
 import ApiHelper from '../../../helpers/ApiHelper'
 import { debounce } from 'lodash';
 
+
 export default class Search extends Component {
   constructor(props) {
     super(props)
