--- studygroups/fixtures/test_courses.json
@@ -1,56 +1,0 @@
-[
-{
-   "fields": {
-      "title": "Programming with Python",
-      "provider": "MIT OCW via P2PU",
-      "link": "http://python.p2pu.org/",
-      "caption": "Slither into computer programming with Python. UniqueCaption1",
-      "on_demand": 1,
-      "topics": "math,science,uniqueTopic1",
-      "created_at": "2015-03-23T15:17:01.880Z",
-      "updated_at": "2015-03-23T15:17:01.880Z"
-   },
-   "model": "studygroups.course",
-   "pk": 1
-},
-{
-   "fields": {
-      "provider": "University of Reading via FutureLearn",
-      "caption": "Get on the write track for university-level writing.",
-      "on_demand": 1,
-      "topics": "writing,uniqueTopic2",
-      "link": "https://www.futurelearn.com/courses/english-for-study",
-      "title": "Academic Writing",
-      "created_at": "2015-03-23T15:18:02.190Z",
-      "updated_at": "2015-03-23T15:17:01.880Z"
-   },
-   "model": "studygroups.course",
-   "pk": 2
-},
-{
-   "fields": {
-      "provider": "University of Washington via Coursera",
-      "caption": "dd",
-      "on_demand": 1,
-      "link": "https://www.coursera.org/learn/public-speaking",
-      "title": "Public Speaking",
-      "created_at": "2015-04-07T20:49:10.718Z",
-      "updated_at": "2015-03-23T15:17:01.880Z"
-   },
-   "model": "studygroups.course",
-   "pk": 3
-},
-{
-   "fields": {
-      "provider": "Khan Academy",
-      "caption": "you don\u2019t have to do it alone. This course will build you.",
-      "on_demand": 1,
-      "link": "https://www.khanacademy.org/math",
-      "title": "GED\u00ae Prep Math",
-      "created_at": "2015-04-07T20:50:04.287Z",
-      "updated_at": "2015-03-23T15:17:01.880Z"
-   },
-   "model": "studygroups.course",
-   "pk": 4
-}
-]
