--- studygroups/fixtures/test_studygroups.json
@@ -1,95 +1,0 @@
-[
-{
-    "fields": {
-      "username": "facilitator",
-      "password": "facilitator123",
-      "email": "facilitator@example.net"
-   },
-   "model": "auth.user",
-   "pk": 1
-},
-{
-   "fields": {
-      "created_at": "2015-03-23T15:18:39.462Z",
-      "updated_at": "2015-03-23T15:18:39.462Z",
-      "name": "The Riders",
-      "course": 3,
-      "start_date": "2015-03-23",
-      "meeting_time": "18:30:00",
-      "end_date": "2015-04-23",
-      "duration": 120,
-      "timezone": "US/Central",
-      "facilitator": 1,
-      "venue_name": "Harold Washington",
-      "venue_address": "123 Street",
-      "venue_details": "3rd floor",
-      "city": "Chicago",
-      "image": "http://image-url.png"
-   },
-   "model": "studygroups.studygroup",
-   "pk": 1
-},
-{
-   "fields": {
-      "created_at": "2015-03-23T15:19:04.318Z",
-      "updated_at": "2015-03-23T15:18:39.462Z",
-      "name": "The Riders",
-      "course": 2,
-      "start_date": "2015-03-23",
-      "meeting_time": "18:00:00",
-      "end_date": "2015-03-23",
-      "duration": 120,
-      "timezone": "US/Central",
-      "facilitator": 1,
-      "venue_name": "Harold Washington",
-      "venue_address": "123 Street",
-      "venue_details": "3rd floor",
-      "city": "Chicago",
-      "image": "http://image-url.png"
-   },
-   "model": "studygroups.studygroup",
-   "pk": 2
-},
-{
-   "fields": {
-      "created_at": "2015-03-25T14:35:02.227Z",
-      "updated_at": "2015-03-23T15:18:39.462Z",
-      "name": "The Red Hawks ",
-      "course": 4,
-      "start_date": "2015-03-23",
-      "meeting_time": "18:30:00",
-      "end_date": "2015-03-23",
-      "duration": 120,
-      "timezone": "US/Central",
-      "facilitator": 1,
-      "venue_name": "Harold Washington",
-      "venue_address": "123 Street",
-      "venue_details": "3rd floor",
-      "city": "Chicago",
-      "image": "http://image-url.png"
-   },
-   "model": "studygroups.studygroup",
-   "pk": 3
-},
-{
-   "fields": {
-      "created_at": "2015-03-25T15:55:44.525Z",
-      "updated_at": "2015-03-23T15:18:39.462Z",
-      "name": "The Bandits of Hell's Bend ",
-      "course": 1,
-      "start_date": "2015-03-23",
-      "meeting_time": "10:00:00",
-      "end_date": "2015-03-23",
-      "duration": 120,
-      "timezone": "US/Central",
-      "facilitator": 1,
-      "venue_name": "Harold Washington",
-      "venue_address": "123 Street",
-      "venue_details": "3rd floor",
-      "city": "Chicago",
-      "image": "http://image-url.png"
-   },
-   "model": "studygroups.studygroup",
-   "pk": 4
-}
-]
